backend/requirements.txt
playwright
beautifulsoup4
lxml
langchain
langchain-openai
langchain-google-genai
langchain-groq
python-dotenv
schedule
supabase
fastapi
uvicorn
deepagents